requirements.txt
schedule==1.2.2
python-dotenv==1.0.1
pyTelegramBotAPI==4.26.0
requests~=2.32.3